changed limits

+++ src/collectives/device/sckl_interpreter.h
@@ -12,7 +12,7 @@
 
 // flags are a 3-tuple of (workindex, gridoffset_iter, step) and it follows a lexicographical order. a threadblock is ahead of another iff its flag is ahead 
 #define COMPUTE_FLAG(__WORKINDEX__,__GRIDOFFSET_ITER__,__STEP__) \
-   SCKL_MAX_ITER*SCKL_MAX_NUM_STEPS*__WORKINDEX__ + (__GRIDOFFSET_ITER__ * SCKL_MAX_NUM_STEPS + __STEP__)
+   SCKL_MAX_ITER*SCKL_MAX_NUM_STEPS*(int64_t)__WORKINDEX__ + ((int64_t)__GRIDOFFSET_ITER__ * SCKL_MAX_NUM_STEPS + (int64_t)__STEP__)
 
 template<typename T, typename PRIMS_WRAPPER>
 class SCKLFunction {
